specs2: tests refactor
