+++ http-testkit-specs2/src/test/scala/com/wix/e2e/http/matchers/drivers/RequestRecorderTestSupport.scala
@@ -1,0 +1,25 @@
+package com.wix.e2e.http.matchers.drivers
+
+import com.wix.e2e.http.HttpRequest
+import com.wix.e2e.http.api.RequestRecordSupport
+import com.wix.e2e.http.matchers.drivers.HttpRequestFactory.aRandomRequest
+import com.wix.e2e.http.matchers.drivers.RequestRecorderFactory.aRequestRecorderWith
+
+trait RequestRecorderTestSupport {
+  val request = aRandomRequest
+  val anotherRequest = aRandomRequest
+  val yetAnotherRequest = aRandomRequest
+  val andAnotherRequest = aRandomRequest
+
+  val anEmptyRequestRecorder = aRequestRecorderWith()
+
+}
+
+object RequestRecorderFactory {
+
+  def aRequestRecorderWith(requests: HttpRequest*) = new RequestRecordSupport {
+    val recordedRequests = requests
+    def clearRecordedRequests() = {}
+  }
+}
+
