+++ http-testkit-specs2/src/test/scala/com/wix/e2e/http/matchers/internal/RequestBodyMatchersTest.scala
@@ -3,7 +3,7 @@ package com.wix.e2e.http.matchers.internal
 import com.wix.e2e.http.matchers.RequestMatchers._
 import com.wix.e2e.http.matchers.drivers.HttpRequestFactory._
 import com.wix.e2e.http.matchers.drivers.MarshallingTestObjects.SomeCaseClass
-import com.wix.e2e.http.matchers.drivers.{CustomMarshallerProvider, HttpResponseTestSupport, MarshallerTestSupport, MatchersTestSupport}
+import com.wix.e2e.http.matchers.drivers.{CustomMarshallerProvider, HttpMessageTestSupport, MarshallerTestSupport, MatchersTestSupport}
 import org.specs2.matcher.ResultMatchers._
 import org.specs2.mutable.Spec
 import org.specs2.specification.Scope
@@ -11,7 +11,7 @@ import org.specs2.specification.Scope
 
 class RequestBodyMatchersTest extends Spec with MatchersTestSupport {
 
-  trait ctx extends Scope with HttpResponseTestSupport with MarshallerTestSupport with CustomMarshallerProvider
+  trait ctx extends Scope with HttpMessageTestSupport with MarshallerTestSupport with CustomMarshallerProvider
 
   "ResponseBodyMatchers" should {
