+++ http-testkit-specs2/src/test/scala/com/wix/e2e/http/matchers/internal/RequestContentTypeMatchersTest.scala
@@ -3,13 +3,13 @@ package com.wix.e2e.http.matchers.internal
 import akka.http.scaladsl.model.ContentTypes._
 import com.wix.e2e.http.matchers.RequestMatchers._
 import com.wix.e2e.http.matchers.drivers.HttpRequestFactory._
-import com.wix.e2e.http.matchers.drivers.{HttpResponseTestSupport, MatchersTestSupport}
+import com.wix.e2e.http.matchers.drivers.{HttpMessageTestSupport, MatchersTestSupport}
 import org.specs2.mutable.Spec
 import org.specs2.specification.Scope
 
 class RequestContentTypeMatchersTest extends Spec with MatchersTestSupport {
 
-  trait ctx extends Scope with HttpResponseTestSupport
+  trait ctx extends Scope with HttpMessageTestSupport
 
   "RequestContentTypeMatchers" should {
