+++ http-testkit-specs2/src/test/scala/com/wix/e2e/http/matchers/internal/RequestCookiesMatchersTest.scala
@@ -1,11 +1,9 @@
 package com.wix.e2e.http.matchers.internal
 
-import akka.http.scaladsl.model.headers.HttpCookiePair
 import com.wix.e2e.http.matchers.RequestMatchers._
+import com.wix.e2e.http.matchers.drivers.CommonTestMatchers._
 import com.wix.e2e.http.matchers.drivers.HttpRequestFactory._
-import com.wix.e2e.http.matchers.drivers.MatchersTestSupport
-import com.wix.test.random._
-import org.specs2.matcher.Matcher
+import com.wix.e2e.http.matchers.drivers.{HttpMessageTestSupport, MatchersTestSupport}
 import org.specs2.matcher.Matchers._
 import org.specs2.mutable.Spec
 import org.specs2.specification.Scope
@@ -13,32 +11,26 @@ import org.specs2.specification.Scope
 
 class RequestCookiesMatchersTest extends Spec with MatchersTestSupport {
 
-  trait ctx extends Scope {
-    val cookie = randomStr -> randomStr
-    val anotherCookie = randomStr -> randomStr
-    val yetAnotherCookie = randomStr -> randomStr
-  }
-
-  def cookieWith(value: String): Matcher[HttpCookiePair] = be_===(value) ^^ { (_: HttpCookiePair).value aka "cookie value" }
+  trait ctx extends Scope with HttpMessageTestSupport
 
   "ResponseCookiesMatchers" should {
 
     "match if cookie with name is found" in new ctx {
-      aRequestWithCookies(cookie) must receivedCookieWith(cookie._1)
+      aRequestWithCookies(cookiePair) must receivedCookieWith(cookiePair._1)
     }
 
     "failure message should describe which cookies are present and which did not match" in new ctx {
-      failureMessageFor(receivedCookieWith(cookie._1), matchedOn = aRequestWithCookies(anotherCookie, yetAnotherCookie)) must
-        contain(cookie._1) and contain(anotherCookie._1) and contain(yetAnotherCookie._1)
+      failureMessageFor(receivedCookieWith(cookiePair._1), matchedOn = aRequestWithCookies(anotherCookiePair, yetAnotherCookiePair)) must
+        contain(cookiePair._1) and contain(anotherCookiePair._1) and contain(yetAnotherCookiePair._1)
     }
 
     "failure message for response withoout cookies will print that the response did not contain any cookies" in new ctx {
-      receivedCookieWith(cookie._1).apply( aRequestWithNoCookies ).message must
+      receivedCookieWith(cookiePair._1).apply( aRequestWithNoCookies ).message must
         contain("Request did not contain any Cookie headers.")
     }
 
     "allow to compose matcher with custom cookie matcher" in new ctx {
-      aRequestWithCookies(cookie) must receivedCookieThat(must = cookieWith(cookie._2))
+      aRequestWithCookies(cookiePair) must receivedCookieThat(must = cookieWith(cookiePair._2))
     }
   }
 }
