+++ http-testkit-specs2/src/test/scala/com/wix/e2e/http/matchers/internal/RequestHeadersMatchersTest.scala
@@ -2,7 +2,7 @@ package com.wix.e2e.http.matchers.internal
 
 import com.wix.e2e.http.matchers.RequestMatchers._
 import com.wix.e2e.http.matchers.drivers.HttpRequestFactory._
-import com.wix.e2e.http.matchers.drivers.{HttpResponseTestSupport, MatchersTestSupport}
+import com.wix.e2e.http.matchers.drivers.{HttpMessageTestSupport, MatchersTestSupport}
 import org.specs2.matcher.AlwaysMatcher
 import org.specs2.mutable.Spec
 import org.specs2.specification.Scope
@@ -10,7 +10,7 @@ import org.specs2.specification.Scope
 
 class RequestHeadersMatchersTest extends Spec with MatchersTestSupport {
 
-  trait ctx extends Scope with HttpResponseTestSupport
+  trait ctx extends Scope with HttpMessageTestSupport
 
   "RequestHeadersMatchers" should {
