+++ http-testkit-specs2/src/test/scala/com/wix/e2e/http/matchers/internal/ResponseContentLengthMatchersTest.scala
@@ -2,14 +2,14 @@ package com.wix.e2e.http.matchers.internal
 
 import com.wix.e2e.http.matchers.ResponseMatchers._
 import com.wix.e2e.http.matchers.drivers.HttpResponseFactory._
-import com.wix.e2e.http.matchers.drivers.{HttpResponseTestSupport, MatchersTestSupport}
+import com.wix.e2e.http.matchers.drivers.{HttpMessageTestSupport, MatchersTestSupport}
 import org.specs2.mutable.Spec
 import org.specs2.specification.Scope
 
 
 class ResponseContentLengthMatchersTest extends Spec with MatchersTestSupport {
 
-  trait ctx extends Scope with HttpResponseTestSupport
+  trait ctx extends Scope with HttpMessageTestSupport
 
   "ResponseContentLengthMatchers" should {
