+++ http-testkit-specs2/src/test/scala/com/wix/e2e/http/matchers/internal/ResponseContentTypeMatchersTest.scala
@@ -2,14 +2,14 @@ package com.wix.e2e.http.matchers.internal
 
 import com.wix.e2e.http.matchers.ResponseMatchers._
 import com.wix.e2e.http.matchers.drivers.HttpResponseFactory._
-import com.wix.e2e.http.matchers.drivers.{HttpResponseTestSupport, MatchersTestSupport}
+import com.wix.e2e.http.matchers.drivers.{HttpMessageTestSupport, MatchersTestSupport}
 import org.specs2.mutable.Spec
 import org.specs2.specification.Scope
 
 
 class ResponseContentTypeMatchersTest extends Spec with MatchersTestSupport {
 
-  trait ctx extends Scope with HttpResponseTestSupport
+  trait ctx extends Scope with HttpMessageTestSupport
 
 
   "ResponseContentTypeMatchers" should {
