+++ http-testkit-specs2/src/test/scala/com/wix/e2e/http/matchers/internal/ResponseCookiesMatchersTest.scala
@@ -3,7 +3,7 @@ package com.wix.e2e.http.matchers.internal
 import com.wix.e2e.http.matchers.ResponseMatchers._
 import com.wix.e2e.http.matchers.drivers.HttpResponseFactory._
 import com.wix.e2e.http.matchers.drivers.HttpResponseMatchers._
-import com.wix.e2e.http.matchers.drivers.{HttpResponseTestSupport, MatchersTestSupport}
+import com.wix.e2e.http.matchers.drivers.{HttpMessageTestSupport, MatchersTestSupport}
 import org.specs2.matcher.Matchers._
 import org.specs2.mutable.Spec
 import org.specs2.specification.Scope
@@ -11,7 +11,7 @@ import org.specs2.specification.Scope
 
 class ResponseCookiesMatchersTest extends Spec with MatchersTestSupport {
 
-  trait ctx extends Scope with HttpResponseTestSupport
+  trait ctx extends Scope with HttpMessageTestSupport
 
   "ResponseCookiesMatchers" should {
