+++ http-testkit-specs2/src/test/scala/com/wix/e2e/http/matchers/internal/ResponseTransferEncodingMatchersTest.scala
@@ -4,14 +4,14 @@ import akka.http.scaladsl.model.TransferEncodings
 import akka.http.scaladsl.model.TransferEncodings._
 import com.wix.e2e.http.matchers.ResponseMatchers._
 import com.wix.e2e.http.matchers.drivers.HttpResponseFactory._
-import com.wix.e2e.http.matchers.drivers.{HttpResponseTestSupport, MatchersTestSupport}
+import com.wix.e2e.http.matchers.drivers.{HttpMessageTestSupport, MatchersTestSupport}
 import org.specs2.mutable.Spec
 import org.specs2.specification.Scope
 
 
 class ResponseTransferEncodingMatchersTest extends Spec with MatchersTestSupport {
 
-  trait ctx extends Scope with HttpResponseTestSupport
+  trait ctx extends Scope with HttpMessageTestSupport
 
 
   "ResponseTransferEncodingMatchersTest" should {
